Se resuelven comentarios para PR
